fix(progress): preserve full quantity range
Two ends of the same inconsistency, both left over from the steps that
put amounts on the screen.

The shortage forms cut what was typed at nine digits. The largest amount
there is has ten, and a task is allowed to need exactly that many, so the
field refused a number the transaction would have taken — and refused it
by dropping a digit rather than by saying so. The limit is gone. What
will not fit is kept as it was written and marked where it was typed, the
way the pipeline boxes already do it.

The three fields that ask for an amount had a parser each and, until
recently, a limit each; that is how the limits came to disagree with one
another and with what a count can be. There is now one account of what an
amount is, and all three read it.

At the other end, a pool section added its tasks up narrow. One task's
total is an Int and the schema keeps it one, but a section holds as many
tasks as the user cares to put in it: measured against real rows through
the real projection, a colour group of three billion came back as
-1294967296 and would have been shown as a section needing a negative
number of pieces. What a section needs and what it still owes are now
counted wide, and counted wide from the first task rather than at the
end. What is counted is pieces; how many tasks there are stays narrow,
because that is bounded by how many rows a person can make.

No schema change: no column, entity or migration moves, and a single
task's own numbers are the width they always were.

+++ app/src/commonMain/kotlin/dev/pnptracker/domain/pools/PoolScreenModel.kt
@@ -67,8 +67,14 @@ data class PoolColorGroup(
     val tasks: List<PoolTask>,
 ) {
     val taskCount: Int get() = tasks.size
-    val requiredTotal: Int get() = tasks.sumOf { it.requiredQuantity ?: 0 }
-    val missingTotal: Int get() = tasks.sumOf { it.currentMissingQuantity }
+
+    // Counted wide, and wide from the first task rather than at the end. One
+    // task's total is an Int and the schema keeps it one, but a group holds as
+    // many tasks as the user cares to put in it, and two large ones already pass
+    // what an Int holds: adding them up narrow and widening the answer showed a
+    // group of three billion as a negative number.
+    val requiredTotal: Long get() = tasks.sumOf { (it.requiredQuantity ?: 0).toLong() }
+    val missingTotal: Long get() = tasks.sumOf { it.currentMissingQuantity.toLong() }
     val failureTotal: Long get() = tasks.sumOf { it.failureTotal }
 }
 
@@ -84,8 +90,12 @@ data class PoolAwaitingColorSection(
     val tasks: List<PoolTask>,
 ) {
     val taskCount: Int get() = tasks.size
-    val requiredTotal: Int get() = tasks.sumOf { it.requiredQuantity ?: 0 }
-    val missingTotal: Int get() = tasks.sumOf { it.currentMissingQuantity }
+
+    // Wide for the same reason a colour group's totals are: this section holds
+    // however many tasks have no colour, and what they add up to is not bounded
+    // by what one of them may be.
+    val requiredTotal: Long get() = tasks.sumOf { (it.requiredQuantity ?: 0).toLong() }
+    val missingTotal: Long get() = tasks.sumOf { it.currentMissingQuantity.toLong() }
     val failureTotal: Long get() = tasks.sumOf { it.failureTotal }
     val isEmpty: Boolean get() = tasks.isEmpty()
 }
